pyspark_project/requirements.txt
# Core PySpark and Big Data
pyspark==3.5.0
pyarrow>=14.0.0
pandas>=2.0.0
numpy>=1.24.0

# Google Cloud Platform
google-cloud-bigquery>=3.11.0
google-cloud-storage>=2.10.0
google-cloud-dataproc>=5.4.0
google-auth>=2.23.0

# Data Processing
pytz>=2023.3
python-dateutil>=2.8.2

# Machine Learning
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Utilities
PyYAML>=6.0
python-dotenv>=1.0.0
tqdm>=4.66.0

# Jupyter and Notebooks
jupyter>=1.0.0
notebook>=7.0.0
ipywidgets>=8.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-spark>=0.6.0

# Code Quality
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0

# Logging and Monitoring
loguru>=0.7.0
